fixed all of that

+++ classes/src/aggregation/ClientsAccountManage.java
@@ -1,5 +1,6 @@
 package aggregation;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface ClientsAccountManage<T> {
@@ -7,6 +8,6 @@ public interface ClientsAccountManage<T> {
     List<T> getNegBalance();
     double getAllCardBalance();
     void sortCardsByBalance();
-    void searchByBalanceInterval(int min,int max);
+    void searchByBalanceInterval(double min, double max);
     void searchByValueType(String type);
 }
